+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Barcode logo round and white
 * Email markup for EventReservation (Barcode needs testing)
 * New settings for add booking and cancel booking pages
+* Barcode logo for Aquatonic
 
 ### 1.2.0: April 7th, 2021
 * Add vendor php-barcode-generator